+++ lava/framework/fencepool.h
@@ -5,6 +5,8 @@
 #ifndef LAVAVK_FENCEPOOL_H
 #define LAVAVK_FENCEPOOL_H
 
+#include <limits>
+#include <numeric>
 #include "lava/framework/object.h"
 #include "core/device.h"
 #include "core/fence.h"
@@ -24,7 +26,7 @@ namespace LavaVk
         FencePool &operator=(FencePool &&) = delete;
 
         Core::SharedFence requestFence(Core::FenceState desiredState = Core::FenceState::Unsignaled);
-        bool wait(uint32_t timeout = std::numeric_limits<uint32_t>::max()) const;
+        bool wait(uint64_t timeout = (std::numeric_limits<uint64_t>::max)()) const;
         void reset();
 
         std::type_index getType() const override;
